Refactor calculation and assembly block in `TickBitmap`
This commit refactors the calculation of mask and the arrangement of assembly block in TickBitmap.sol for clarity and easier maintainability. Now, the bitPos utilizes a consistent hex representation across the file. The bitPos overflow scenario is handled taking into consideration that 1 << 256 equals 0.

+++ src/SqrtPriceMath.sol
@@ -93,7 +93,7 @@ library SqrtPriceMath {
                     : FullMath.mulDiv(amount, FixedPoint96.Q96, liquidity)
             );
 
-            nextSqrtPrice = (sqrtPX96 + quotient).toUint160();
+            nextSqrtPrice = (uint256(sqrtPX96) + quotient).toUint160();
         } else {
             uint256 quotient = (
                 amount <= type(uint160).max
